+++ src/jtl/Connector/Shopware/Mapper/CustomerOrder.php
@@ -32,7 +32,7 @@ class CustomerOrder extends DataMapper
 
     public function findStatus($id)
     {
-        return (intval($id) == 0) ? null : $this->Manager()->getRepository('Shopware\Models\Order\Status')->find($id);
+        return $this->Manager()->getRepository('Shopware\Models\Order\Status')->find($id);
     }
 
     public function findAll($limit = 100, $count = false, $from = null, $until = null)