+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+2.0.12
+------
+- Fixed customer push salutation bug
+- Fixed customer order status push
+- Added new product attribute sw_pseudo_sales
+- Added paypal plus installment
+
 2.0.11
 ------
 - Added customer mapping via email